staging: apf: Fix address calculation for SG-DMA transfers
This patch comes from:
  https://github.com/Xilinx/linux-xlnx.git

If a buffer allocated with the use of sds_alloc is used in an
SG-DMA transfer such that only the end of the buffer is
transferred, the address calculation needed to skip the beginning
of the buffer was incorrect.  This patch fixes that issue.

Signed-off-by: Michael Gill <gill@xilinx.com>
(cherry picked from commit bbf9d396cae64edec762990701326d37c5ec90d1)
Signed-off-by: Zumeng Chen <zumeng.chen@windriver.com>