usb: pci-quirks: add Intel USB drcfg mux device
commit 581950b59293ec6023218b024d54cfa6415fd319 from
git://git.yoctoproject.org/linux-yocto-4.1

In some Intel platforms, a single usb port is shared between USB host
and device controllers. The shared port is under control of a switch
which is defined in the Intel vendor defined extended capability for
xHCI.

This patch adds the support to detect and create the platform device
for the port mux switch.

Signed-off-by: Lu Baolu <baolu.lu@linux.intel.com>
Reviewed-by: Felipe Balbi <balbi@kernel.org>
Signed-off-by: Wan Ahmad Zainie <wan.ahmad.zainie.wan.mohamad@intel.com>
Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>
[Yongxin Liu: Remove pset since struct property_set has been deleted in
include/linux/property.h]
Integrated-by: Yongxin Liu <yongxin.liu@windriver.com>