mmc: omap_hsmmc: support non-1.8V IO controllers
This patch comes from:
  git://git.ti.com/processor-sdk/processor-sdk-linux.git

OMAP HSMMC driver assumes that if the controller does
not support dual-volt, then it supports only 1.8V IO.

This assumption can be incorrect. For example, on K2G
MMC0 supports 3.3V IO only. AM57x Beagle-x15 and IDK
boards support only 3.3V IO on eMMC interface.

Support device-tree property "no-1-8-v" to for
controllers which are not dual-voltage and do not
support 1.8V IO.

Note that lack of support for this property has not
led to any known regression in affected platforms so
far. But it will be nice to be in sync with hardware
configuration.

Signed-off-by: Sekhar Nori <nsekhar@ti.com>
(cherry picked from commit 15445a7408274d2eed6c088e1f1a987b123bdad4)
Signed-off-by: Zumeng Chen <zumeng.chen@windriver.com>