hid: intel-ish-hid: add warm_reset flag for ishtp_bus_remove_all_clients
commit 746f80624c499af08db0a4a756126f7266447d10 from
https://github.com/01org/linux-apollolake-i

ishtp_bus_remove_all_clients() is called in ish_remove() and
ishtp_reset_handler(). Warm_reset argument/flag is added to identify which
caller function calls to ishtp_bus_remove_all_clients(). When firmware
resets, the client devices, which are being used, are not required to be
removed. However, they are removed when intel-ish-ipc module is
unloaded/removed.

Signed-off-by: Ooi, Joyce <joyce.ooi@intel.com>
Signed-off-by: Kevin Hao <kexin.hao@windriver.com>