drm/vmwgfx: Replace numeric parameter like 0444 with macro
commit 2d8e60e8b0742b7a5cddc806fe38bb81ee876c33 upstream

I find that the developers often just specified the numeric value
when calling a macro which is defined with a parameter for access
permission.

As we know, these numeric value for access permission have had the
corresponding macro, and that using macro can improve the robustness and
readability of the code, thus, I suggest replacing the numeric parameter
with the macro.

Signed-off-by: Chuansheng Liu <chuansheng.liu@intel.com>
Signed-off-by: Baole Ni <baolex.ni@intel.com>
Reviewed-by: Sinclair Yeh <syeh@vmware.com>
Signed-off-by: Sinclair Yeh <syeh@vmware.com>