rpmb: enable emmc specific read data fixup
commit b5f8ee2b7e6b2d0d582e76eff62b437f466c286e from
git://git@github.com:01org/linux-intel-4.9.git

For eMMC the block count of the RPMB read operation is not indicated in
the original RPMB Data Read Request packet.
This might be different then the implementation of other protocol
standards.
This patch implements a fixup for this behavior.

V6: new in the series
V7: resend
V8: resend

Change-Id: I2b24e43d9f205fe1cdbf37cfdd9b307097d2ad38
Tracked-On: https://jira01.devtools.intel.com/browse/OAM-46382
Signed-off-by: Tomas Winkler <tomas.winkler@intel.com>
Signed-off-by: Alexander Usyskin <alexander.usyskin@intel.com>
Tested-by: Avri Altman <avri.altman@sandisk.com>