net: macb: add support for mdio phy nodes
This patch comes from:
  https://github.com/Xilinx/linux-xlnx.git

This patch adds support for mdio phy nodes.

With this patch the macb driver first tries to find
the mdio node.
If it is available will create the phy/mdio devices for the
phy/mdio nodes available in the mdio.

If the mdio node is not available it will try to probe the phy nodes
available in the mac nodes as the driver does earlier.

Signed-off-by: Kedareswara rao Appana <appanad@xilinx.com>
Signed-off-by: Soren Brinkmann <soren.brinkmann@xilinx.com>
(cherry picked from commit f9c43e8386e46b129761b724d0268298e755bb36)
Signed-off-by: Zumeng Chen <zumeng.chen@windriver.com>