clk: zynqmp: Add zynqmp ultrascale gate support
This patch comes from:
  https://github.com/Xilinx/linux-xlnx.git

Add zynqmp ultrascale gate support
Based on the common gate with the IO accessors customized This is
needed so as to not break multi-platform support.
Also spinlocking is removed as it is taken care by firmware.

Signed-off-by: Shubhrajyoti Datta <shubhrajyoti.datta@xilinx.com>
Signed-off-by: Michal Simek <michal.simek@xilinx.com>
(cherry picked from commit 85926715c882e8c662505a0419685546f751add3)
Signed-off-by: Zumeng Chen <zumeng.chen@windriver.com>