phy: core: add API to support phy reset
This patch comes from:
  git://git.ti.com/processor-sdk/processor-sdk-linux.git

In certain situations, such as after a remote device
is restarted, a PHY needs to be reset in order to
regain synchronization with the remote device.

The reset does not necessarily mean a complete restart
and/or reconfigurration of the PHY.  It may just be the
toggling of one bit in one register or the reconfiguration
of a small subset of the registers of the PHY.

This patch adds a new API to support the reset of a
PHY in these situations.

Signed-off-by: WingMan Kwok <w-kwok2@ti.com>
Signed-off-by: Sekhar Nori <nsekhar@ti.com>
(cherry picked from commit 8b3d9976f401e3c6628434eef6ec49d0a170cb7b)
Signed-off-by: Zumeng Chen <zumeng.chen@windriver.com>