ASoC: Intel: Skylake: Update module id in pin connections
commit 200673d63cfbd3f08bebe913a000460331fcdbf6 from
git://git@github.com:01org/linux-intel-4.9.git

Currently, each module instance id is getting read from
firmware binary and updated in the driver. However, the
module id in the statically mapped pins of a module is
still the same as coming from topology binary. This causes
module binding failure if the module id is wrong in the
topology binary. This patch updates the module id in the
pin connections of each module instance.

Signed-off-by: Panwar, Ashish <ashish.panwar@intel.com>
Signed-off-by: Jeeja KP <jeeja.kp@intel.com>
[Kevin: Just some minor context mods in order to port to wrlinux]
Signed-off-by: Kevin Hao <kexin.hao@windriver.com>