drm: modify drm_global_item_ref to avoid two times of writing ref->object
commit d28d6e6fa55e08a16a83ae572b1cbf4f03e8fbf1 from
https://github.com/Xilinx/linux-xlnx.git

In previous drm_global_item_ref, there are two times of writing
ref->object if item->refcount is 0. So this patch does a minor update
to put alloc and init ref firstly, and then to modify the item of glob
array. Use "else" to avoid two times of writing ref->object. It can
make the code logic more clearly.

Reviewed-by: Christian König <christian.koenig@amd.com>
Signed-off-by: Huang Rui <ray.huang@amd.com>
Signed-off-by: Alex Deucher <alexander.deucher@amd.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>