perf scripts: Fixing file name extension
Original patch taken from: https://github.com/Linaro/OpenCSD.git

With this patch it is possible to handle library files that
don't end with ".so".  Something like ".so.0.1" is now supported
and will add the proper entry in the mmaps dictionary.

(Modification to the regular expression was suggested by
Tor Jeremiassen)

Signed-off-by: Mathieu Poirier <mathieu.poirier@linaro.org>
Signed-off-by: Jianchuan Wang <jianchuan.wang@windriver.com>