kgdboc, usb: Add polling support for kgdboc and usb serial
Add polling support to the USB serial driver and kgdboc.

For the USB polling support to work the usb interrupt routine is
called while the kernel is stopped in kgdb.  This will be called with
interrupts disabled as well as all of the CPUs being halted.  This
might not work so well on an SMP system...

Signed-off-by: Jason Wessel <jason.wessel@windriver.com>
Signed-off-by: He Zhe <zhe.he@windriver.com>

Fixed serial_poll_init(), serial_poll_get_char(),
serial_poll_put_char() in drivers/usb/serial/usb-serial.c

Signed-off-by: Kexin(Casey) Chen <Casey.Chen@windriver.com>