modules,kgdb: Add module gone event
When a debugger such as gdb is attached, it cares when a kernel module
is either in the coming state or the gone state such that it can
update the debugger state for mapping symbols to actual code that is
loaded by the kernel.  The module "GOING" state is not good enough
because the kernel will still list the module as being in the section
maps and occupying virtual memory.

Signed-off-by: Jason Wessel <jason.wessel@windriver.com>
CC: Mark Asselstine <mark.asselstine@windriver.com>

Adjust hunk #1 of kernel/module.c

Signed-off-by: He Zhe <zhe.he@windriver.com>

Signed-off-by: Kexin(Casey) Chen <Casey.Chen@windriver.com>