net: ethernet: ti: allow cpts to be built separately
This patch comes from:
  git://git.ti.com/processor-sdk/processor-sdk-linux.git

TI CPTS IP is used as part of TI OMAP CPSW driver, but it's also
present as part of NETCP on TI Keystone 2 SoCs. So, It's required
to enable build of CPTS for both this drivers and this can be
achieved by allowing CPTS to be built separately.

Hence, allow cpts to be built separately and convert it to be
a module as both CPSW and NETCP drives can be built as modules.

Signed-off-by: Grygorii Strashko <grygorii.strashko@ti.com>
Signed-off-by: Sekhar Nori <nsekhar@ti.com>
(cherry picked from commit 81b993867582c07f02152e2a2de32616e40acd60)
Signed-off-by: Zumeng Chen <zumeng.chen@windriver.com>