staging: fsl-dpaa2: Allocate extra headroom in skb
The Linux IP forwarding code compares the socket buffer headroom to a
16-byte aligned headroom reported by the netdevice. If the socket buffer
does not accommodate this value, the stack will make a reallocation for
each frame, thus impacting performance.

Allocate an extra headroom to mitigate this issue. The value is chosen
based on the options set in the buffer layout, the Rx buffer alignment
value and the netdevice required headroom.

This brings a ~10% performance increase in IPv4 forwarding scenarios
involving linear skbs. SG skbs don't benefit from this patch since their
headroom is much smaller than the netdev's requirements, and the
performance gain would be smaller due to the much lower fps.

Signed-off-by: Bogdan Purcareata <bogdan.purcareata@nxp.com>
[Original patch taken from SDK-V2.0-1703]
Signed-off-by: Meng Li <Meng.Li@windriver.com>