x86/PCI: VMD: Move VMD driver to drivers/pci/host
commit 181ffd19cc9898f795646bf9a897072a419a51ed upstream.

Move the driver source and Kconfig to the PCI host bridge drivers directory
and move the config option to a more appropriate sub-menu instead of
occupying the top-level location.

Update the Kconfig option with the X86_64 dependency that was implicitly
included from the previous location, and add information about the module
name when built as a loadable module.

Signed-off-by: Keith Busch <keith.busch@intel.com>
Signed-off-by: Bjorn Helgaas <bhelgaas@google.com>
CC: Jon Derrick <jonathan.derrick@intel.com>
Signed-off-by: Liwei Song <liwei.song@windriver.com>