HACK: remoteproc/pru: extend DT-based GP_MUX_SEL programming to K2G ICE EVM
The support to configure a PRU core's GP_MUX_SEL value in the PRUSS_GPCFG
register from the DT property "ti,pruss-gp-mux-mode" has been extended for
the K2G ICE EVM board as well. This configuration is needed on K2G ICE EVM
board to allow the PRU Ethernet pins to be routed properly from each PRUSS
to the two PHYs supported by a PRUSS. This configuration was previously
provided only to the AM571x-IDK board in commit a67dccf15475 ("HACK:
remoteproc/pru: allow device tree to configure GP_MUX_SEL").

The generalization of this feature is to be evaluated in the future to
determine whether this configuration needs to be done by a client driver
through an API or to be done purely from DT.

Reported-by: Roger Quadros <rogerq@ti.com>
Signed-off-by: Suman Anna <s-anna@ti.com>
(cherry picked from commit 89960f200f78933181ac58bd53c40380c3d2a320)