dpa_offload/dpa_stats: Fix bad propagation of counters offsets to US during sync get_counters
The user space driver expects some updates in the counters Ids array.
It expects to see there, for the user space counters, the exact offsets
where it needs to fill in the statistics data. This data was propagated
in case of async calls to dpa_stats_get_counters, but NOT in case of
synchronous calls.

Signed-off-by: Marian Chereji <marian.chereji@nxp.com>
[Original patch taken from SDK-V2.0-1703]
Signed-off-by: Meng Li <Meng.Li@windriver.com>