staging: fsl-mc: Add function to set obj_id to destroy
The patch fixes an error reproducible when setting the object id to
destroy. This affects restool scenarios where objects are detroyed.

Add a single function in the MC command header to set the object id in a
destroy command. Rename the object offset and size macros, since the
values ar applicable for destroy as well.

Signed-off-by: Bogdan Purcareata <bogdan.purcareata@nxp.com>
[Original patch taken from SDK-V2.0-1703]
Signed-off-by: Meng Li <Meng.Li@windriver.com>