ASoC: Intel: Skylake: Add support for LPMode
commit 4316f15b94233dffd446dc47bea8ad55a25d7d58 from
git://git@github.com:01org/linux-intel-4.9.git

For D0i3, we need to tell DSP to run the pipelines in LP mode. This
information is kept in topology and passed to driver as an attribute
for pipe.

So add a new tuple for lpmode and program the pipe based on value set.

Signed-off-by: Jayachandran B <jayachandran.b@intel.com>
Signed-off-by: Vinod Koul <vinod.koul@intel.com>
Signed-off-by: Mark Brown <broonie@kernel.org>