net: phy: broadcom: Add BCM54810 PHY entry
commit b14995ac2527b43a75c9190fbd4efd43fb1f4562 upstream

The BCM54810 PHY requires some semi-unique configuration, which results
in some additional configuration in addition to the standard config.
Also, some users of the BCM54810 require the PHY lanes to be swapped.
Since there is no way to detect this, add a device tree query to see if
it is applicable.

Inspired-by: Vikas Soni <vsoni@broadcom.com>
Signed-off-by: Jon Mason <jon.mason@broadcom.com>
Reviewed-by: Florian Fainelli <f.fainelli@gmail.com>
Reviewed-by: Andrew Lunn <andrew@lunn.ch>
Signed-off-by: David S. Miller <davem@davemloft.net>

Conflicts:
	drivers/net/phy/Kconfig
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>