drm/nouveau/kms/nv50: separate out base/ovly channel usage bounds commit
commit 6bbab3b6b65669b974bece5a688b3dba49ea3e06 upstream

This commit separates the calculation of EVO state from the commit, in
order to make the same code useful for atomic modesetting.

The legacy interfaces have been wrapped on top of them.

We're no longer touching the overlay channel usage bounds as of this
commit.  The code to do so is in place for when overlay planes are
added.

Signed-off-by: Ben Skeggs <bskeggs@redhat.com>