drm/nouveau/kms: prepare to support suspend/resume of display state with atomic
commit b167db0e68dca4f4009c1b07270ce9bb509c86cd upstream

This is different from the equivilant functions in the atomic helpers in
that we fully disable the pipe instead of just setting it to inactive.

We do this (primarily) to ensure the framebuffer cleanup paths are hit,
allowing buffers to be un-pinned from memory so they can be evicted to
system memory and not lose their contents while suspended.

Signed-off-by: Ben Skeggs <bskeggs@redhat.com>