drm/nouveau/fifo/gf100-: provide notification to user if channel is killed
commit ff9f29abf0ef4c43e696bef7621884518e6bdbda upstream

There are instances (such as non-recoverable GPU page faults) where
NVKM decides that a channel's context is no longer viable, and will
be removed from the runlist.

This commit notifies the owner of the channel when this happens, so
it has the opportunity to take some kind of recovery action instead
of hanging.

Signed-off-by: Ben Skeggs <bskeggs@redhat.com>