drm: xilinx: dp_sub: Store drm formats
commit  8e0dd86c0a715c52d094efcd9deb5921f429e301 from
https://github.com/Xilinx/linux-xlnx.git

Store all supported formats for each layer. Stored formats are returned
to support dynamic format switching. With this, it's not required to
specify the layer format in DT. Thus, the layer pixel format is optional,
and it falls back to the format format. Still user can specify the layer
to be specific format, for example for fbdev emulation.

Signed-off-by: Hyun Kwon <hyun.kwon@xilinx.com>
Signed-off-by: Michal Simek <michal.simek@xilinx.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>