net: netcp: fix K2L/E system lockup caused by gbe cable unplug
This patch comes from:
  git://git.ti.com/processor-sdk/processor-sdk-linux.git

On K2E/L, if ethernet cable is unplugged when the gbe switch is
sending packets to the network, it can cause a system lockup.

The issue is created when the GMII_EN bit is cleared while a
packet is in transmission. This resets the MAC state machine.
When the GMII_EN is then set back to a ‘1’, the MAC state machine
detects that the FIFO is not at start of frame and proceeds to
purge the rest of the frame.

When the end of frame is detected, it issues a tx_stat_valid,
but since the MAC state machine was not in the transmit states,
there is no status (good, collisions , etc). When the stat block
sees the tx_stat_valid it goes to the transmit state and since
there is no status there was no exit condition. Then the next
stats read will hang the bus.

This patch provides a software workaround that avoids the lockup
by disabling the port statistics when the link goes down. When
the link comes back up, the MAC is reset and port statistics is
re-enabled when the MAC becomes idle after reset.

Signed-off-by: WingMan Kwok <w-kwok2@ti.com>
Signed-off-by: Sekhar Nori <nsekhar@ti.com>
(cherry picked from commit a1ff83583f7b25e5bfb2b9e8629b515837ba6d85)
Signed-off-by: Zumeng Chen <zumeng.chen@windriver.com>